Completed file re-organization suggestion on the PR comments. Moved all serializers to their respective files (e.g. business_serializers.py has the serializer for class business, etc.), and moved view functions to appropriate files for better organization. helper_views.py contain helper view functions that aren't necessarily used directly in the code but I used for testing if things were being added to the database correctly, and they might be useful later. Modified some urls.py files to contain correct import statement since files were reorganized.
