+++ backend/pages/serializers/genre_serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from pages.models import Genre
+
+# Serializer for genres
+class GenreSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Genre
+        fields = ['id', 'genre']
